refactor(chat): split ai-bubble.parts into separate microcomponents files
